Refactor: REST API 리팩토링
