+++ KHS/.gitignore
@@ -1,3 +1,2 @@
-app.js
 node_modules/
 postcss.config.js
